Affected pages:
http://editor.dart-lang.appspot.com/downloads.html
http://editor.dart-lang.appspot.com/docs/editor/
http://editor.dart-lang.appspot.com/docs/editor/getting-started/
http://editor.dart-lang.appspot.com/docs/editor/troubleshoot.html

While I was in the editor/ page, I gave the settings section a clearer title and a TOC entry, and I moved it to the end. I also fixed some random "win32" tags.

Seth, could you review the code?

I picked Jaime as the content reviewer in part because of time zone benefits, but feel free to pass the buck to someone else!

Background: We used to point to only the 64-bit download for Mac. As the site got beautified, the 32-bit download came back and the link to the continuous build became much harder to find. Also, some windows-only information started being shown all the time.

Review URL: https://chromiumcodereview.appspot.com/11259002