book update: map -> mappedBy; html event handling and httprequest changes
The main changes are here:
http://book.dart-lang.appspot.com/docs/dart-up-and-running/contents/ch03.html#ch03-httprequest

Also see the summary of updates, so far:
http://book.dart-lang.appspot.com/docs/dart-up-and-running/contents/preface.html#preface-updates

Ignore the changes to contents/index.html. That file is mostly hidden, and the changes appear to be spurious (maybe due to a remote stylesheet adding links where none used to be before).

Review URL: https://chromiumcodereview.appspot.com/12226066