remove no-op/unhelpful dart_analyzer flags from dartlang test script
It turns out that --enable-type-checks and --type-checks-for-inferred-types are
no-ops left over from when this was a real compiler (dartc). --incremental is
similarly meaningless; it might do something, but it probably isn't what we want,
and it doesn't affect execution time (a little under 6 minutes on my Mac).

Review URL: https://chromiumcodereview.appspot.com/13959003