Embed DartPad in the top part of homepage.
You have to click to see the DartPad. That's intentional, to avoid
possible delays.

Better ideas for the code are welcome, but I'd rather publish and iterate
than wait until the code is ideal. We intend to update the rest of the code
on the page in future changes, and probably use embedded DartPad in more
places on the homepage.

Staged:
https://homepage-newdp-dot-dart-lang.appspot.com/

R=shaza@google.com

Review URL: https://chromiumcodereview.appspot.com/1319523005