Effective Dart!
Harvested the existing "best practice" guides and re-organized them into
a coherent guide broken into a few pages. This replaces:

- articles/style-guide
- articles/api-naming-guide
- articles/doc-comment-guidelines
- articles/idiomatic-dart

And creates:

- effective-dart
- effective-dart/style
- effective-dart/documentation
- effective-dart/usage
- effective-dart/design

There are a few new guidelines here, and a couple of changed ones, but
this is mostly just rearranging things and refreshing the prose.

I'm planning to do a couple of follow-up changes to fill in more
guidelines.

R=kathyw@google.com

Review URL: https://chromiumcodereview.appspot.com/1399193004 .